@@ -12,15 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/animations": "^8.0.1",
+    "@angular/common": "^8.0.1",
+    "@angular/compiler": "^8.0.1",
+    "@angular/core": "^8.0.1",
+    "@angular/forms": "^8.0.1",
+    "@angular/http": "8.0.0-beta.9",
+    "@angular/platform-browser": "^8.0.1",
+    "@angular/platform-browser-dynamic": "^8.0.1",
+    "@angular/router": "^8.0.1",
     "core-js": "^2.6.8",
     "front-markjs": "^8.9.1-1",
     "ngx-bootstrap": "^4.2.0",
@@ -29,10 +29,10 @@
     "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "^0.800.3",
+    "@angular/cli": "^8.0.1",
+    "@angular/compiler-cli": "^8.0.1",
+    "@angular/language-service": "^8.0.1",
     "@types/jasmine": "^3.3.12",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^12.0.2",
@@ -53,7 +53,7 @@
     "showdown": "^1.9.0",
     "ts-node": "^8.2.0",
     "tslint": "^5.16.0",
-    "typescript": "3.2.4",
+    "typescript": "3.4.5",
     "wordcount": "^1.1.1"
   }
 }
